LPS-55475 Move the classes to the proper packages and changes the template to generate the code accordingly
