+++ modules/util/portal-tools-service-builder/src/main/resources/com/liferay/portal/tools/service/builder/dependencies/persistence_impl.ftl
@@ -74,8 +74,8 @@ import com.liferay.portal.security.auth.PrincipalThreadLocal;
 import com.liferay.portal.security.permission.InlineSQLHelperUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.ServiceContextThreadLocal;
+import com.liferay.portal.service.persistence.CompanyProvider;
 import com.liferay.portal.service.persistence.impl.BasePersistenceImpl;
-import com.liferay.portal.service.persistence.impl.CompanyProvider;
 import com.liferay.portal.service.persistence.impl.NestedSetsTreeManager;
 import com.liferay.portal.service.persistence.impl.PersistenceNestedSetsTreeManager;
 import com.liferay.portal.service.persistence.impl.TableMapper;
@@ -1716,12 +1716,20 @@ public class ${entity.name}PersistenceImpl extends BasePersistenceImpl<${entity.
 	}
 
 	<#if entity.isShardedModel()>
-		@BeanReference(type = CompanyProvider.class)
+		<#if osgiModule>
+			@ServiceReference(type = CompanyProvider.class)
+		<#else>
+			@BeanReference(type = CompanyProvider.class)
+		</#if>
 		protected CompanyProvider companyProvider;
 	<#else>
 		<#list entity.columnList as column>
 			<#if column.isCollection() && column.isMappingManyToMany()>
-				@BeanReference(type = CompanyProvider.class)
+				<#if osgiModule>
+					@ServiceReference(type = CompanyProvider.class)
+				<#else>
+					@BeanReference(type = CompanyProvider.class)
+				</#if>
 				protected CompanyProvider companyProvider;
 
 				<#break>
