+++ portal-impl/src/META-INF/company-provider-spring.xml
@@ -7,9 +7,9 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
 >
-	<bean id="com.liferay.portal.service.persistence.impl.CompanyProvider" class="com.liferay.portal.service.persistence.impl.ThreadLocalCompanyProvider" />
+	<bean id="com.liferay.portal.service.persistence.CompanyProvider" class="com.liferay.portal.service.persistence.impl.ThreadLocalCompanyProvider" />
 
-	<bean id="com.liferay.portal.service.persistence.impl.CompanyProviderHolder" class="com.liferay.portal.service.persistence.impl.CompanyProviderHolderImpl">
-		<property name="companyProvider" ref="com.liferay.portal.service.persistence.impl.CompanyProvider" />
+	<bean id="com.liferay.portal.service.persistence.CompanyProviderHolder" class="com.liferay.portal.service.persistence.impl.CompanyProviderHolderImpl">
+		<property name="companyProvider" ref="com.liferay.portal.service.persistence.CompanyProvider" />
 	</bean>
 </beans>
